Add verbose printing to sandbox.
The sandbox prints to stdout making it hard to redirect generic output
from the results JSON output.  It now prints some messages (the ones
getting in the way for the moment) to whatever verbose_log is set to
(stderr by default in run-tron-game.py (which will come in a future
commit)).